refactor card entity split task-card entity with a base card entity
